9 months after update, now with 24.5 Mojo
This change fixes compatibility issues with current stable Mojo release.
Performance is largely similar, with minor degradations here and there,
and some improvements in other places (where I was finally able to
turn on controlled loop unrolls properly).

Some bugs are gone, some are still there, filed two more in the process.
New bugs are crashes, so at least they are easier to spot, but the old
silent ones are also now peskier than ever it seems.

I've toyed briefly with the nightly builds and it seems 24.6 will change
the syntax further. Oh joy.

Overall I suppose '24 is still not the year for production usage of this thing.